\name{delt-package}
\alias{delt-package}
\alias{delt}
\docType{package}
\title{
Estimation of multivariate densities with adaptive histograms
}
\description{
The package implements methods for estimating multivariate
densities: adaptive histograms (greedy histograms and
CART-histograms), stagewise minimization, and bootstrap
aggregation are provided.
}
\details{
\tabular{ll}{
Package: \tab delt\cr
Version: \tab 0.6.0\cr
Date: \tab 2006-05-14\cr
Depends: \tab R\cr
License: \tab GPL version 2 or newer\cr
URL: \tab http://www.rni.helsinki.fi/~jsk/delt\cr
Packaged: \tab Wed May 17 10:47:27 2006; jsk\cr
Built: \tab R 2.2.1; i686-pc-linux-gnu; 2006-05-17 10:47:34; unix\cr
}

Greedy histograms:
eval.greedy,
lstseq.greedy.

CART-histograms:
eval.cart,
lstseq.cart.

CART histograms step by step:
densplit,
prune,
eval.pick.

Bootstrap aggregation of histograms:
eval.bagg,
lstseq.bagg.

Stagewise minimization:
eval.stage,
eval.stage.gauss.

Other utilities:
partition,
plotparti.

Tree transformation:
lefrig2par,
makebina.

Miscallenous:
intpcf,
supp.

Index:
\preformatted{
densplit                Calculation of an overfitting histogram
eval.bagg               Returns a bootstrap aggregation of adaptive
                        histograms
eval.cart               Calculates a CART histogram
eval.greedy             Returns a greedy histogram
eval.pick               Returns a subtree of an evaluation tree
eval.stage              Returns a stagewise minimization estimate
eval.stage.gauss        Returns a 1D Gaussian mixture density estimate
intpcf                  Calculates the integral of a piecewise constant
                        function
lefrig2par              Transforms an evaluation tree so that it can be
                        plotted with the "plottree" function of package
                        "denpro"
lstseq.bagg             Calculates a scale of bootstrap aggregated
                        histograms
lstseq.cart             Calculates a scale of CART histograms
lstseq.greedy           Calculates a scale of greedy histograms
makebina                Tranforms and evaluation tree to the tree
                        object of R
partition               Finds the partition generated by an evaluation
                        tree
plotparti               Draws a partition
prune                   Prepares for pruning an overfitting evaluation
                        tree
scaspa                  Finds the number of modes of histograms which
                        are obtained by pruning an overfitting
                        histogram
supp                    Returns the bounding box of observations
}


}
\author{
Jussi Klemelä <klemela@rumms.uni-mannheim.de>

Maintainer: Jussi Klemelä <klemela@rumms.uni-mannheim.de>
}
%\references{}
\keyword{ smooth }
%\seealso{~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~ }
\examples{

# Generate the data

dendat<-sim.data(n=500,seed=5,type="mulmodII")

# Calculate the estimates

eva<-eval.greedy(dendat,leaf=16)

eva<-eval.cart(dendat,leaf=16)

eva<-eval.bagg(dendat,B=3,leaf=12,prune="on")

eva<-eval.stage(dendat,leaf=10,M=3)

# Draw the estimates

lst<-leafsfirst(eva)
plotvolu(lst)

dp<-draw.pcf(eva,pnum=c(60,60))         
persp(dp$x,dp$y,dp$z,theta=-20,phi=30)

}



